+++ src/common/utils/composeValidators.js
@@ -1,3 +1,3 @@
-export const composeValidators = (...validators) => (value) => {
-  return validators.reduce((error, validator) => error || validator(value), undefined);
+export const composeValidators = (...validators) => (...data) => {
+  return validators.reduce((error, validator) => error || validator(...data), undefined);
 };