+++ src/components/BatchAllocations/hooks/useBatchAllocationFormatter/useBatchAllocationFormatter.js
@@ -17,7 +17,7 @@ import {
   AmountWithCurrencyField,
   FieldSelectFinal,
   FieldTags,
-  validateRequiredNotNegative,
+  validateRequired,
 } from '@folio/stripes-acq-components';
 
 import { composeValidators } from '../../../../common/utils';
@@ -32,7 +32,9 @@ import {
   validateAllocationAfterField,
   validateBudgetStatus,
   validateFundStatus,
+  validateNotNegative,
   validateNumericValue,
+  validateResponseError,
 } from './validators';
 
 import css from './styles.css';
@@ -91,7 +93,11 @@ export const useBatchAllocationFormatter = (intl, fiscalYear) => {
           fullWidth
           marginBottom0
           name={`${FINANCE_DATA}.${item[ROW_INDEX]}.${BATCH_ALLOCATION_FIELDS.fundStatus}`}
-          validate={validateFundStatus(intl)}
+          validate={composeValidators(
+            validateRequired,
+            validateFundStatus(intl),
+            validateResponseError(),
+          )}
           validateFields={[]}
         />
       );
@@ -116,7 +122,11 @@ export const useBatchAllocationFormatter = (intl, fiscalYear) => {
           fullWidth
           marginBottom0
           name={`${FINANCE_DATA}.${item[ROW_INDEX]}.${BATCH_ALLOCATION_FIELDS.budgetStatus}`}
-          validate={validateBudgetStatus(intl)}
+          validate={composeValidators(
+            validateRequired,
+            validateBudgetStatus(intl),
+            validateResponseError(),
+          )}
           validateFields={[]}
         />
       );
@@ -131,7 +141,10 @@ export const useBatchAllocationFormatter = (intl, fiscalYear) => {
           name={`${FINANCE_DATA}.${item[ROW_INDEX]}.${BATCH_ALLOCATION_FIELDS.budgetAllocationChange}`}
           required
           type="number"
-          validate={validateNumericValue(intl)}
+          validate={composeValidators(
+            validateNumericValue(intl),
+            validateResponseError(),
+          )}
           validateFields={[]}
         />
       );
@@ -146,6 +159,7 @@ export const useBatchAllocationFormatter = (intl, fiscalYear) => {
           marginBottom0
           name={`${CALCULATED_FINANCE_DATA}.${item[ROW_INDEX]}.${BATCH_ALLOCATION_FIELDS.budgetAfterAllocation}`}
           required
+          type="number"
           validate={validateAllocationAfterField(intl, item[ROW_INDEX])}
           validateFields={[]}
         />
@@ -161,7 +175,11 @@ export const useBatchAllocationFormatter = (intl, fiscalYear) => {
           name={`${FINANCE_DATA}.${item[ROW_INDEX]}.${BATCH_ALLOCATION_FIELDS.budgetAllowableEncumbrance}`}
           required
           type="number"
-          validate={composeValidators(validateNumericValue(intl), validateRequiredNotNegative)}
+          validate={composeValidators(
+            validateNumericValue(intl),
+            validateNotNegative(intl),
+            validateResponseError(),
+          )}
           validateFields={[]}
         />
       );
@@ -176,7 +194,11 @@ export const useBatchAllocationFormatter = (intl, fiscalYear) => {
           name={`${FINANCE_DATA}.${item[ROW_INDEX]}.${BATCH_ALLOCATION_FIELDS.budgetAllowableExpenditure}`}
           required
           type="number"
-          validate={composeValidators(validateNumericValue(intl), validateRequiredNotNegative)}
+          validate={composeValidators(
+            validateNumericValue(intl),
+            validateNotNegative(intl),
+            validateResponseError(),
+          )}
           validateFields={[]}
         />
       );
@@ -192,6 +214,7 @@ export const useBatchAllocationFormatter = (intl, fiscalYear) => {
           parse={identity}
           placeholder="Description"
           type="text"
+          validate={validateResponseError()}
           validateFields={[]}
         />
       );
@@ -206,11 +229,12 @@ export const useBatchAllocationFormatter = (intl, fiscalYear) => {
             labelless
             marginBottom0
             name={`${FINANCE_DATA}.${item[ROW_INDEX]}.transactionTag.tagList`}
+            validate={validateResponseError()}
           />
         </div>
       );
     },
-  }), [form, intl, fundStatusOptions, fiscalYear?.currency, budgetStatusOptions]);
+  }), [budgetStatusOptions, fiscalYear?.currency, form, fundStatusOptions, intl]);
 
   return formatter;
 };